[BUGFIX] Prevent "Nesting level too deep" error in search
To prevent duplicates in the search result, there is a check via
`in_array` in place. The default behavior of `in_array` is a comparison
of all properties of the objects. This can result in a PHP Fatal error
"Nesting level too deep".

The error can be prevented by using a strict comparison, which can be
achieved by passing `true` as third parameter.